+++ test/diagnostics/init.lua
@@ -640,6 +640,46 @@ end
 ]]
 
 TEST [[
+local x, <!y!>, <!z!> = 1
+print(x, y, z)
+]]
+
+TEST [[
+local x, y, <!z!> = 1, 2
+print(x, y, z)
+]]
+
+TEST [[
+local x, y, z = print()
+print(x, y, z)
+]]
+
+TEST [[
+local x, y, z
+print(x, y, z)
+]]
+
+TEST [[
+local x, y, z
+x, <!y!>, <!z!> = 1
+print(x, y, z)
+]]
+
+TEST [[
+X, <!Y!>, <!Z!> = 1
+]]
+
+TEST [[
+T = {}
+T.x, <!T.y!>, <!T.z!> = 1
+]]
+
+TEST [[
+T = {}
+T['x'], <!T['y']!>, <!T['z']!> = 1
+]]
+
+TEST [[
 ---@class <!Class!>
 ---@class <!Class!>
 ]]